Move component to shared directory and change name
