--- frontend/src/app/features/wdf/wdf-data-change/wdf-requirements-state/wdf-requirements-state.component.ts
@@ -1,12 +1,0 @@
-import { Component, Input, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-wdf-requirements-state',
-  templateUrl: './wdf-requirements-state.component.html',
-})
-export class WdfRequirementsStateComponent implements OnInit {
-  @Input() overallWdfEligibility: boolean;
-  @Input() currentWdfEligibility: boolean;
-
-  ngOnInit(): void {}
-}
