+++ frontend/src/app/shared/components/funding-requirements-state/funding-requirements-state.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-funding-requirements-state',
+  templateUrl: './funding-requirements-state.component.html',
+})
+export class FundingRequirementsStateComponent implements OnInit {
+  @Input() overallWdfEligibility: boolean;
+  @Input() currentWdfEligibility: boolean;
+  @Input() singeStaffRecord: boolean;
+
+  ngOnInit(): void {}
+}
